refactor(texts): rename folder
