+++ src/modules/texts/Text/components/CardText/CardText.tsx
@@ -2,7 +2,7 @@ import { useTheme } from "@mui/material";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import React, { memo, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import CardTextView from "./CardTextView";
+import CardTextView from "modules/texts/Text/components/CardText/CardTextView";
 import { mountBodyText } from "modules/texts/Text/utils/mount-body-text";
 import { createURLQueryParams } from "shared/util/createURLQueryParams";
 import { TEXT_QUERY_KEY } from "shared/services/constants/texts";
@@ -10,8 +10,8 @@ import { useDialogText } from "modules/texts/Text/shared/hooks/useDialogText";
 import { useContextText } from "modules/texts/Text/Context";
 import { fetchPostTextService } from "modules/texts/Text/service";
 import { IFetchPostTextResponse } from "modules/texts/Text/types";
-import { fetchGetAllTextsService } from "modules/texts/Text/components/CardText/service";
-import { IFetchGetAllTextsResponse } from "modules/texts/Text/components/CardText/service/types";
+import { fetchGetAllTextsService } from "modules/texts/Text/components/CardText/services";
+import { IFetchGetAllTextsResponse } from "modules/texts/Text/components/CardText/services/types";
 
 const CardText: React.FC = () => {
 	const { palette } = useTheme();
