+++ src/modules/texts/Text/components/CardText/services/fetchGetAllTextsService.ts
@@ -1,5 +1,5 @@
 import api from "shared/services/api";
-import { IFetchGetAllTextsResponse } from "modules/texts/Text/components/CardText/service/types";
+import { IFetchGetAllTextsResponse } from "modules/texts/Text/components/CardText/services/types";
 
 export const fetchGetAllTextsService = async (boardID: string | undefined) => {
 	const { data } = await api.get<IFetchGetAllTextsResponse>(
