+++ src/modules/texts/Text/components/CardText/services/index.ts
