+++ src/modules/texts/Text/components/CardText/services/types/index.ts
